+++ packages/frontend/src/components/DataViz/transformers/useChart.ts
@@ -4,6 +4,7 @@ import {
     isVizCartesianChartConfig,
     isVizPieChartConfig,
     PieChartDataModel,
+    type PivotChartData,
     type VizCartesianChartConfig,
     type VizPieChartConfig,
 } from '@lightdash/common';
@@ -17,16 +18,27 @@ export const useChart = <T extends ResultsRunner>({
     sql,
     projectUuid,
     limit,
+    orgColors,
+    onPivot,
+    slug,
+    uuid,
 }: {
     config?: VizCartesianChartConfig | VizPieChartConfig;
     resultsRunner: T;
     sql?: string;
     projectUuid?: string;
     limit?: number;
+    orgColors?: string[];
+    onPivot?: (pivotData: PivotChartData) => void;
+    slug?: string;
+    uuid?: string;
 }) => {
     const chartTransformer = useMemo(() => {
         if (config?.type === ChartKind.PIE) {
-            return new PieChartDataModel({ resultsRunner, config });
+            return new PieChartDataModel({
+                resultsRunner,
+                fieldConfig: config?.fieldConfig,
+            });
         }
         if (
             config?.type === ChartKind.VERTICAL_BAR ||
@@ -34,10 +46,10 @@ export const useChart = <T extends ResultsRunner>({
         ) {
             return new CartesianChartDataModel({
                 resultsRunner,
-                config,
+                fieldConfig: config.fieldConfig,
             });
         }
-    }, [resultsRunner, config]);
+    }, [resultsRunner, config?.fieldConfig, config?.type]);
 
     const getTransformedData = useCallback(
         async () =>
@@ -46,13 +58,21 @@ export const useChart = <T extends ResultsRunner>({
                 sql,
                 projectUuid,
                 limit,
+                slug,
+                uuid,
             ),
         // TODO: FIX THIS ISSUE - it should include the SQL, but the sql shouldn't change on change, but on run query
         // eslint-disable-next-line react-hooks/exhaustive-deps
         [chartTransformer, config?.fieldConfig, projectUuid, limit],
     );
 
-    const transformedData = useAsync(getTransformedData, [getTransformedData]);
+    const transformedData = useAsync(async () => {
+        const data = await getTransformedData();
+        if (onPivot && data) {
+            onPivot(data);
+        }
+        return data;
+    }, [getTransformedData]);
 
     const chartSpec = useMemo(() => {
         if (!transformedData.value) return undefined;
@@ -74,9 +94,10 @@ export const useChart = <T extends ResultsRunner>({
                 transformedData.value,
                 config.display,
                 config.type,
+                orgColors,
             );
         }
-    }, [chartTransformer, config, transformedData.value]);
+    }, [chartTransformer, config, transformedData.value, orgColors]);
 
     return {
         ...transformedData,