+++ packages/frontend/src/features/sqlRunner/components/Header/HeaderEdit.tsx
@@ -95,8 +95,8 @@ export const HeaderEdit: FC = () => {
                                 description={
                                     savedSqlChart.description ?? undefined
                                 }
-                                viewStats={1} // todo: update endpoint to return view stats
-                                firstViewedAt={undefined}
+                                viewStats={savedSqlChart.views}
+                                firstViewedAt={savedSqlChart.firstViewedAt}
                                 withChartData={false}
                             />
                         </Group>